platform/polyglot/app_yaml_static_config/py/pyproject.toml
[project]
name = "app_yaml_static_config"
version = "0.1.0"
description = "Polyglot YAML Configuration Loading, Storage, and Retrieval"
requires-python = ">=3.11,<3.12"
dependencies = [
    "PyYAML==6.0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app_yaml_static_config"]

[tool.pytest.ini_options]
testpaths = ["__tests__"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

